+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Crash occurring when changing/submitting input concurrently to game over.
+
 ## [1.10.0] - 2022-09-21
 
 ### Added